Replace Jest with mocha (and chai, jsdom, sinon)
