+++ package.json
@@ -14,15 +14,21 @@
     "react-input-autosize": "^0.4.3"
   },
   "devDependencies": {
+    "babel": "^5.5.6",
     "babel-eslint": "^3.1.1",
-    "babel-jest": "^5.2.0",
+    "chai": "^3.0.0",
     "eslint": "^0.22.1",
     "eslint-plugin-react": "^2.2.0",
     "gulp": "^3.8.11",
-    "jest-cli": "^0.4.2",
+    "istanbul": "^0.3.15",
+    "jsdom": "^3.1.2",
     "lessify": "^1.0.1",
+    "mocha": "^2.2.5",
+    "mocha-jsdom": "^0.4.0",
     "react": ">=0.12.0",
-    "react-component-gulp-tasks": "^0.7.0"
+    "react-component-gulp-tasks": "^0.7.0",
+    "sinon": "^1.15.1",
+    "sinon-chai": "^2.8.0"
   },
   "peerDependencies": {
     "react": ">=0.12.0"
@@ -46,7 +52,8 @@
     "examples": "gulp dev:server",
     "lint": "eslint .",
     "release": "gulp release",
-    "test": "jest; true",
+    "test": "mocha --compilers js:babel/register",
+    "cover": "istanbul cover node_modules/.bin/_mocha -- -u exports --compilers js:babel/register -R spec",
     "watch": "gulp watch:lib"
   },
   "jest": {
